+++ app/src/main/java/com/example/chatapplication/Chat_Fragment.java
@@ -2,7 +2,9 @@ package com.example.chatapplication;
 
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import android.view.LayoutInflater;
@@ -14,9 +16,19 @@ import android.widget.TextView;
 import com.example.chatapplication.Model.disp_msgModel;
 import com.example.chatapplication.RecyclerViewMessages.MessagesAdapter;
 import com.google.android.material.textfield.TextInputLayout;
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+
+import es.dmoral.toasty.Toasty;
 
 public class Chat_Fragment extends Fragment {
 
@@ -26,7 +38,7 @@ public class Chat_Fragment extends Fragment {
 
     private String mParam1;
     private String mParam2;
-    private String email;
+    private String userToChatWith;
 
     List<disp_msgModel> dispMsgModelList;
 
@@ -41,8 +53,8 @@ public class Chat_Fragment extends Fragment {
 
     }
 
-    public Chat_Fragment(String email) {
-        this.email=email;
+    public Chat_Fragment(String userToChatWith) {
+        this.userToChatWith=userToChatWith;
     }
 
     public static Chat_Fragment newInstance(String param1, String param2) {
@@ -68,6 +80,70 @@ public class Chat_Fragment extends Fragment {
                              Bundle savedInstanceState) {
 
         View view= inflater.inflate(R.layout.fragment_chat, container, false);
+        send_btn=view.findViewById(R.id.sendMsg_btn);
+        msg=view.findViewById(R.id.msg);
+        userChattingHead=view.findViewById(R.id.head);
+        currentUser= FirebaseAuth.getInstance().getCurrentUser();
+
+        userChattingHead.setText(userToChatWith);
+
+        chatsRec=view.findViewById(R.id.chatmsgRec);
+        chatsRec.setLayoutManager(new LinearLayoutManager(view.getContext()));
+        chatsRec.setHasFixedSize(true);
+
+
+        send_btn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                String message=msg.getEditText().getText().toString();
+                if(message.isEmpty())
+                    Toasty.warning(view.getContext(),"You cannot send a empty message",Toasty.LENGTH_SHORT).show();
+                else{
+                    sendMsg(currentUser.getEmail(),userToChatWith,message);
+                    msg.getEditText().setText("");
+                    Toasty.success(view.getContext(),"Message sent successfully",Toasty.LENGTH_SHORT).show();
+                }
+            }
+        });
+        readMsg(currentUser.getEmail(),userToChatWith);
         return view;
+    }
+
+    //working fine
+    private void sendMsg(String sender,String receiver,String msg) {
+        Map<String,Object> mapObj=new HashMap<>();
+        mapObj.put("sender",sender);
+        mapObj.put("receiver",receiver);
+        mapObj.put("msg",msg);
+
+        FirebaseDatabase.getInstance().getReference().child("Chats").push().setValue(mapObj);
+    }
+
+    private void readMsg(String me, String him){
+        dispMsgModelList=new ArrayList<>();
+
+        FirebaseDatabase.getInstance().getReference("Chats")
+                .addValueEventListener(new ValueEventListener() {
+                    //everytime data is added,it checks whether it ito be added to the List for displaying messages or not using a for each loop
+                    @Override
+                    public void onDataChange(@NonNull DataSnapshot parentSnapshot) {
+                        dispMsgModelList.clear();
+                        for (DataSnapshot childSnapshot:parentSnapshot.getChildren()){
+                            disp_msgModel chatM=childSnapshot.getValue(disp_msgModel.class);
+                            if(chatM.getSender().equals(me) && chatM.getReceiver().equals(him)||
+                                    chatM.getSender().equals(him) && chatM.getReceiver().equals(me))
+                            {
+                                dispMsgModelList.add(chatM);
+                            }
+                        }
+                    }
+
+                    @Override
+                    public void onCancelled(@NonNull DatabaseError error) {
+
+                    }
+                });
+        messagesAdapter=new MessagesAdapter(dispMsgModelList);
+        chatsRec.setAdapter(messagesAdapter);
     }
 }
