+++ app/src/main/java/com/example/chatapplication/RecyclerViewActiveUsers/ActiveUsersAdapter.java
@@ -1,21 +1,15 @@
 package com.example.chatapplication.RecyclerViewActiveUsers;
 
 import android.content.Context;
-import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.bumptech.glide.Glide;
-import com.example.chatapplication.ActiveUsersFragment;
-import com.example.chatapplication.ChatPage;
 import com.example.chatapplication.Chat_Fragment;
 import com.example.chatapplication.Model.userModel;
 import com.example.chatapplication.R;
