--- app/src/main/res/layout/chat_page.xml
@@ -1,1 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
