+++ app/src/main/res/layout/fragment_active_users.xml
@@ -5,7 +5,9 @@
     android:layout_height="match_parent"
     tools:context=".ActiveUsersFragment">
 
-    <!-- TODO: Update blank fragment layout -->
+    <RelativeLayout
+        android:layout_height="match_parent"
+        android:layout_width="match_parent">
     <Button
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -19,5 +21,5 @@
         android:layout_height="match_parent"
         android:id="@+id/activeUsers"
         />
-
+    </RelativeLayout>
 </FrameLayout>
